refactor: Change column bool tests from e2e to integration tests.
