--- tests/serverpod_test_server/lib/src/endpoints/column_operations/column_bool.dart
@@ -1,65 +1,0 @@
-import 'package:serverpod/serverpod.dart';
-import 'package:serverpod_test_server/src/generated/protocol.dart';
-
-class ColumnBoolEndpoint extends Endpoint {
-  Future<void> insert(Session session, List<Types> types) async {
-    for (var type in types) {
-      await Types.db.insertRow(session, type);
-    }
-  }
-
-  Future<int> deleteAll(Session session) async {
-    var result =
-        await Types.db.deleteWhere(session, where: (_) => Constant.bool(true));
-    return result.length;
-  }
-
-  Future<List<Types>> findAll(Session session) async {
-    return await Types.db.find(
-      session,
-      where: (_) => Constant.bool(true),
-    );
-  }
-
-  Future<List<Types>> equals(Session session, bool? value) async {
-    return await Types.db.find(
-      session,
-      where: (t) => t.aBool.equals(value),
-    );
-  }
-
-  Future<List<Types>> notEquals(Session session, bool? value) async {
-    return await Types.db.find(
-      session,
-      where: (t) => t.aBool.notEquals(value),
-    );
-  }
-
-  Future<List<Types>> inSet(Session session, List<bool> value) async {
-    return await Types.db.find(
-      session,
-      where: (t) => t.aBool.inSet(value.toSet()),
-    );
-  }
-
-  Future<List<Types>> notInSet(Session session, List<bool> value) async {
-    return await Types.db.find(
-      session,
-      where: (t) => t.aBool.notInSet(value.toSet()),
-    );
-  }
-
-  Future<List<Types>> isDistinctFrom(Session session, bool value) async {
-    return await Types.db.find(
-      session,
-      where: (t) => t.aBool.isDistinctFrom(value),
-    );
-  }
-
-  Future<List<Types>> isNotDistinctFrom(Session session, bool value) async {
-    return await Types.db.find(
-      session,
-      where: (t) => t.aBool.isNotDistinctFrom(value),
-    );
-  }
-}
